i#15: claim tag conventions
There are two different claim tag conventions in use; in future
the best strategy may be to use one convention for AP devices
and the other (which is already in use) for non-AP devices.

Add a comment to csregisters.h explaining this, and define two
new pairs of claim tag macros.